refactor: 관리자용 이용권 컨트롤러 분리
